chore: reformat and reorganize file structure
