+++ app/App.js
@@ -10,10 +10,10 @@ import {
 } from 'react-native';
 
 //Component imports
-import { HomeScreen } from './Home'
-import { AllPartners } from './AllPartners'
-import { Message } from './Message'
-import { User } from './User'
+import { HomeScreen } from './components/Home'
+import { AllPartners } from './components/AllPartners'
+import { Message } from './components/Message'
+import { User } from './components/User'
 
 export const ClimbingPartnerFinder = StackNavigator({
   Home: { screen: HomeScreen },
